backend/requirements/local.txt
-r ./base.txt

# Debugging
ipdb==0.13.2

# Tools
django-extensions==2.2.9

# Testing
mypy==0.770
pytest==5.4.2
pytest-sugar==0.9.3
pytest-django==3.9.0
factory-boy==2.12.0
